requirements.txt
# mysqlclient==2.2.7
Django==5.2
djangorestframework==3.16.0
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
PyJWT==2.10.1
